Move scripts around.
