+++ packages/vital-elements/package.json
@@ -30,8 +30,8 @@
     "copy": "copyfiles -u 1 \"./src/*.css\" \"./lib/\"",
     "clean": "rimraf \"lib/*\" && rimraf tsconfig.tsbuildinfo && rimraf \"doc/api\"",
     "tokens:clean": "rimraf \"src/generated/*\"",
-    "tokens:find": "node ./lib/cjs/scripts/bin/find-tokens.js",
-    "tokens:parse": "node ./lib/cjs/scripts/bin/parse-tokens.js"
+    "tokens:find": "node ./lib/cjs/scripts/find-tokens.js",
+    "tokens:parse": "node ./lib/cjs/scripts/parse-tokens.js"
   },
   "dependencies": {
     "@bodiless/cli": "^1.0.0-rc.42",
