+++ packages/vital-elements/src/scripts/parseTokens/FigmaVariableName.ts
